% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_compare_crosstabs.R
\name{compare_crosstabs}
\alias{compare_crosstabs}
\title{Utility: Compare crosstabs across datasets}
\usage{
compare_crosstabs(orig_data, new_data, vars, k = 2)
}
\arguments{
\item{orig_data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object)
containing the original data.}

\item{new_data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object)
containing the data from imputation or synthesis.}

\item{vars}{A character vector of categorical variables' names
to compared.}

\item{k}{An integer between 2 and the number of variables in \code{vars},
controlling the dimension of the cross-tabulations.
For example, specifying \code{k=3} will produce all three-way cross-tabulations from the variables
specified by \code{vars}.}
}
\value{
A list of data frames representing cross-tabulations.
In each data frame, the variable \code{N_ORIG} and \code{N_NEW} give the counts
from the original and new datasets, respectively.
}
\description{
Computes k-way cross-tabulations
in the original versus new data.
}
\examples{
# Create an example variable to use

library(dplyr)

libraries_orig <- libraries_orig |> mutate(
  STAFF_SIZE_CATEGORY = TOTSTAFF |> cut(
    breaks = c(0, 1.262, 3.3, 7.34, 17.868, Inf),
    labels = c("Small", "Small-to-Medium", "Medium",
               "Medium-to-Large", "Large"),
  ),
  HAS_ELECTRONIC_MATERIALS = ifelse(ELMATCIR > 0, "Yes", "No")
)

libraries_synth <- libraries_synth |> mutate(
  STAFF_SIZE_CATEGORY = TOTSTAFF |> cut(
    breaks = c(0, 1.262, 3.3, 7.34, 17.868, Inf),
    labels = c("Small", "Small-to-Medium", "Medium",
               "Medium-to-Large", "Large"),
  ),
  HAS_ELECTRONIC_MATERIALS = ifelse(ELMATCIR > 0, "Yes", "No")
)

# Compare crosstabs in the two datasets
compare_crosstabs(
  orig_data = libraries_orig,
  new_data  = libraries_synth,
  vars      = c("STABR", "STAFF_SIZE_CATEGORY",
                "HAS_ELECTRONIC_MATERIALS"),
  k         = 2 
)

}
