% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_pairwise_associations.R
\name{compare_pairwise_correlations}
\alias{compare_pairwise_correlations}
\title{Utility: compare pairwise correlations}
\usage{
compare_pairwise_correlations(orig_data, new_data, vars, method = "pearson")
}
\arguments{
\item{orig_data}{A data frame (or tbl) corresponding
to the original data.}

\item{new_data}{A data frame (or tbl) containing the
imputed or synthetic data.}

\item{vars}{A character vector of variables to be analyzed.}

\item{method}{Type of correlation: either \code{"pearson"} (the default)
or \code{"spearman"}.}
}
\value{
A data frame with one row per pair of variables specified by \code{vars}.
The columns \code{ORIG_CORR} and \code{NEW_CORR} give the correlations
in the original and new datasets, respectively. The column \code{DIFF}
gives the difference (\code{NEW_CORR} minus \code{ORIG_CORR}).
The column \code{REL_DIFF} gives the relative difference
(\code{DIFF} divided by \code{ORIG_CORR}).
}
\description{
Computes correlations
for all pairs of variables, separately
in the original data and in the new data.
Computes differences in the correlation values
between the original and new data.
}
\examples{
compare_pairwise_correlations(
  orig_data = libraries_orig,
  new_data  = libraries_synth,
  vars      = c("TOTCIR", "ELMATCIR", "LIBRARIA")
)
}
