% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_ci_overlap.R
\name{compute_io}
\alias{compute_io}
\title{Internal helper function to compute interval overlap statistic}
\usage{
compute_io(l1, u1, l2, u2)
}
\arguments{
\item{l1}{The lower limit of the first confidence interval}

\item{u1}{The upper limit of the first confidence interval}

\item{l2}{The lower limit of the second confidence interval}

\item{u2}{The upper limit of the second confidence interval}
}
\description{
Computes interval overlap statistic
}
\keyword{internal}
