% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_distribution_distances.R
\name{compute_ks_distance}
\alias{compute_ks_distance}
\title{Utility measure: Kolmogorov-Smirnov distance for continuous variables}
\usage{
compute_ks_distance(orig_data, new_data, vars)
}
\arguments{
\item{orig_data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object)
containing the original data.}

\item{new_data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object)
containing the data from imputation or synthesis.}

\item{vars}{A character vector of categorical variables' names
to compared.}
}
\value{
A data frame with one row per variable analyzed.
The column "KS_DIST" gives the Kolmogorv-Smirnov distance.
The column "P_VALUE" gives the two-sided p-value for the
asymptotic two-sample Kolmogorov-Smirnov test. The columns
"N_ORIG" and "N_NEW" give the total number of observations
with non-missing values for the variable in the
in the original and new datasets, respectively.
}
\description{
Computes measures of distance
between the distributions of continuous variables
in the original versus new data.
}
\details{
Computes the Kolmogorv-Smirnov distance between the
distributions of the variable in the two datasets.
This is simply the largest difference between the empirical
cumulative distribution functions of the two datasets.

Also computes the two-sided p-value for the asymptotic
two-sample Kolmogorv-Smirnov test. See \code{\link[stats:ks.test]{stats::ks.test()}}
for more details.
}
\examples{

# Example of distances between categorical variables

compute_ks_distance(
  orig_data    = libraries_orig,
  new_data     = libraries_synth,
  vars         = c("LIBRARIA", "TOTSTAFF")
)
}
