% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_u_statistic.R
\name{compute_u_statistic}
\alias{compute_u_statistic}
\title{Compute U-statistic}
\usage{
compute_u_statistic(orig_data, new_data, vars = NULL)
}
\arguments{
\item{orig_data}{A data frame (or tbl) corresponding
to the original data.}

\item{new_data}{A data frame (or tbl) containing the
imputed or synthetic data.}

\item{vars}{A character vector of variables
to use as predictors in the logistic regression model}
}
\value{
A data frame with the U-statistic, its standard error, and p-value
}
\description{
Computes the U-statistic
based on a logistic regression model with specific predictor variables.
See Snoke et al. (2018) for details.
}
\examples{
compute_u_statistic(
  orig_data = libraries_orig,
  new_data  = libraries_synth,
  vars      = c("TOTCIR", "ELMATCIR")
)
}
\references{
Joshua Snoke, Gillian M. Raab, Beata Nowok, Chris Dibben, Aleksandra Slavkovic.
"General and Specific Utility Measures for Synthetic Data."
Journal of the Royal Statistical Society Series A: Statistics in Society,
Volume 181, Issue 3, June 2018, Pages 663–688.
https://doi.org/10.1111/rssa.12358
}
