% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_distribution_distances.R
\name{dist_btwn_cat_vars}
\alias{dist_btwn_cat_vars}
\title{Utility measure: distance for categorical variables}
\usage{
dist_btwn_cat_vars(orig_data, new_data, vars, wgt = NULL, new_suffixes = NULL)
}
\arguments{
\item{orig_data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object)
containing the original data.}

\item{new_data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object)
containing the data from imputation or synthesis.}

\item{vars}{A character vector of categorical variables' names
to compared.}

\item{wgt}{An optional character string giving the name
of a weight variable (which must be present in both datasets)
to use for comparing weighted distributions.}

\item{new_suffixes}{An optional character vector
giving the suffixes used to denote different
implicates in \code{new_data}.}
}
\value{
A data frame with one row per variable analyzed,
and columns \code{"HD"} and \code{"W_HD"} giving
Hellinger's distance for the unweighted distributions
and weighted distributions, respectively.
}
\description{
Computes measures of distance
between the distributions of categorical variables
in the original versus new data. Currently only supports
Hellinger's distance.
}
\details{
Hellinger's distance is computed for a variable \eqn{X}
with \eqn{k} categories. The distance depends on the
vectors of estimated proportions in two datasets,
denoted \eqn{P = (p_1,\dots,p_k)} and \eqn{Q = (q_1,\dots,q_k)}.
The distance is defined as follows:
\deqn{
  HD = \frac{1}{\sqrt{2}} \sqrt{\sum_{i=1}^{k} \left(\sqrt{p_i} - \sqrt{q_i}\right)^2}
}
Other software packages (e.g., the R package 'dad')
sometimes omit the term \eqn{\frac{1}{\sqrt(2)}}.

If a weights variable is specified, then the weighted Hellinger's distance (\code{W_HD})
is computed using the formula above, but applied to proportions estimated using the weights.
}
\examples{

# Example of distances between categorical variables

orig_data <- data.frame(
  'RACETHM' = sample(1:4, size = 100, replace = TRUE),
  'WGT'     = runif(n = 100, 10, 15)
)

new_data <- data.frame(
  'RACETHM' = sample(1:4, size = 100, replace = TRUE),
  'WGT'     = runif(n = 100, 10, 15)
)

dist_btwn_cat_vars(
  orig_data    = orig_data,
  new_data     = new_data,
  vars         = "RACETHM",
  wgt          = "WGT"
)

# Example when the synthetic data has multiple implicates
new_data_multiple_implicates <- data.frame(
  'RACETHM_1' = sample(1:4, size = 100, replace = TRUE),
  'RACETHM_2' = sample(1:4, size = 100, replace = TRUE),
  'RACETHM_3' = sample(1:4, size = 100, replace = TRUE),
  'RACETHM_4' = sample(1:4, size = 100, replace = TRUE),
  'RACETHM_5' = sample(1:4, size = 100, replace = TRUE),
  'WGT'       = runif(n = 100, min = 10, max = 15)
)

dist_btwn_cat_vars(
  orig_data    = orig_data,
  new_data     = new_data_multiple_implicates,
  vars         = "RACETHM",
  wgt          = "WGT",
  new_suffixes = paste0("_", 1:5)
)

}
