% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_distribution_distances.R
\name{dist_btwn_cont_vars}
\alias{dist_btwn_cont_vars}
\title{Utility: distances between continuous variables' distributions}
\usage{
dist_btwn_cont_vars(data, vars, wgt = NULL, new_suffixes, og_suffix)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object).}

\item{vars}{Character strings giving the name of variables in the data}

\item{wgt}{A character string giving the name of a weight variable in the data}

\item{new_suffixes}{A character vector giving the suffixes
used to distinguish imputed/synthetic variables from the original variables.
For example, with three implicates these might be \code{c("_1", "_2", "_3")}.}

\item{og_suffix}{A character string giving the suffix for every original variable
(the suffix must be shared by all of the variables in \code{vars}).}
}
\value{
A data frame with one column per summary statistic,
and one row per combination of variable and implicate.
}
\description{
Summaries of the distances between original variables in the data
and their imputed or synthetic versions
}
\examples{

data('libraries')

dist_btwn_cont_vars(
  orig_data = libraries_orig,
  new_data  = libraries_synth,
  vars      = c("TOTCIR", "ELMATCIR"),
  statistic = "mean"
)

}
