% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_distribution_distances.R
\name{HD}
\alias{HD}
\title{Internal helper function to compute Hellinger's distance}
\usage{
HD(data, p_var, q_var)
}
\arguments{
\item{data}{A data frame with one row
per category of a variable of interest (e.g., PhD field).}

\item{p_var}{A character string giving the name
of the variable containing an empirical distribution
function's values (e.g., sample percentages from original data).}

\item{q_var}{A character string giving the name
of a different variable containing an alternative
empirical distribution function's values (
e.g., sample percentages from synthetic data).}
}
\value{
A data frame with the column named 'HD',
giving Hellinger's Distance
}
\description{
Helper function to compute
Hellinger's Distance given a data frame
containing two different empirical distributions
}
\keyword{internal}
