% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_distribution_distances.R
\name{summ_stats_var}
\alias{summ_stats_var}
\title{Internal helper function to compute weighted summary statistics}
\usage{
summ_stats_var(data, var, new_suff, og_suff)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g., a \code{tbl_spark} object).}

\item{var}{A character string giving the name of a variable in the data}

\item{new_suff}{A suffix used to distinguish the imputed variable from the original variable}

\item{og_suff}{The suffix for the original variable}
}
\value{
A data frame with one column per summary statistic
}
\description{
Summarize the absolute differences between
a variable and its imputed or synthetic version
}
\examples{

example_data <- data.frame(
  'SALARY_15' = rnorm(n = 10, mean = 75000, sd = 5000),
  'SALARY_15_1' = rnorm(n = 10, mean = 75000, sd = 5000)
)

summ_stats_var(
  data     = example_data, 
  var      = "SALARY_15", 
  new_suff = "_1", 
  og_suff  = ""
)

}
\keyword{internal}
