% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_ci_overlap.R
\name{svy_check_ci_overlap}
\alias{svy_check_ci_overlap}
\title{Check confidence interval overlap for survey data}
\usage{
svy_check_ci_overlap(
  orig_svy,
  new_svy,
  vars,
  new_suffixes = NULL,
  statistic = "mean",
  level = 0.95
)
}
\arguments{
\item{orig_svy}{A survey design object corresponding
to the original data.}

\item{new_svy}{A survey design object containing the
imputed or synthetic data.}

\item{vars}{A character vector of variables to be analyzed.}

\item{new_suffixes}{An optional character vector of suffixes
used to distinguish multiple implicates.}

\item{statistic}{Either \code{"mean"}, \code{"total"}, or \code{"prop"}.}

\item{level}{The confidence level for the intervals. Default is 0.95.
Must be at least 0.5 and less than 1.}
}
\value{
A data frame with one row per combination of variable and implicate.
The column \code{OVERLAP} is the confidence interval overlap statistic,
while \code{NEW_EST_IN_ORIG_CI} indicates whether the point estimate
for the new data is contained in the confidence interval from the original data.
The remaining columns are the point estimates and confidence intervals
from each source.
}
\description{
Compare overlap in confidence intervals
for a statistic computed using original survey data vs
imputed/synthetic survey data.
}
\details{
This method was proposed by Karr et al. (2006); see section 2.2.1, particularly equation (5)
of that reference.
}
\examples{
library(srvyr)

# Prepare survey data for analysis
libraries_orig_svy <- as_survey_rep(
  .data      = libraries_orig,
  weights    = FULL_SAMPLE_WGT,
  repweights = num_range("REP_WGT_", 1:80),
  type       = "successive-difference", mse = TRUE
)
libraries_synth_svy <- as_survey_rep(
  .data      = libraries_synth,
  weights    = FULL_SAMPLE_WGT,
  repweights = num_range("REP_WGT_", 1:80),
  type       = "successive-difference", mse = TRUE
)

# Check CI overlap for estimated means
svy_check_ci_overlap(
  orig_svy  = libraries_orig_svy,
  new_svy   = libraries_synth_svy,
  vars      = c("TOTCIR", "ELMATCIR"),
  statistic = "mean"
)
}
\references{
Karr, A. F., Kohnen, C. N., Oganian, A., Reiter, J. P., & Sanil, A. P. (2006).
"A Framework for Evaluating the Utility of Data Altered to Protect Confidentiality."
The American Statistician, 60(3), 224–232.
https://doi.org/10.1198/000313006x124640
}
\seealso{
\code{\link[=check_ci_overlap]{check_ci_overlap()}} for use with non-survey data.
}
