% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthassess-package.R
\docType{package}
\name{synthassess-package}
\alias{synthassess}
\alias{synthassess-package}
\title{synthassess: Utility and Risk Assessments for Synthetic Data}
\description{
Conducts utility and risk assessments for synthetic data.
}
\author{
\strong{Maintainer}: Benjamin Schneider \email{BenjaminSchneider@westat.com}

Authors:
\itemize{
  \item Robyn Ferg \email{RobynFerg@westat.com}
}

}
\keyword{internal}
